requirements.txt
torch>=2.0
torchvision>=0.15
facenet-pytorch>=2.5.3
opencv-python>=4.8
Pillow>=9.0
numpy>=1.24
face_recognition>=1.3
